issue_794: unit tests for proxy range feature were updated
